tests/wpt/tests/svg/painting/parsing/marker-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing marker with invalid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#MarkerProperty"/>
    <h:meta name="assert" content="marker supports only the paint grammar 'none | marker-ref'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_invalid_value("marker", "auto");
test_invalid_value("marker", 'none url("https://example.com/")');

  ]]></script>
</svg>
